requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
apscheduler==3.10.4
httpx==0.26.0
pydantic==2.6.0
python-dotenv==1.0.0
stripe==7.0.0
python-telegram-bot==20.7
